small fixes

+++ src/main/scala/ru/org/codingteam/keter/game/objects/equipment/items/Weapon.scala
@@ -3,9 +3,9 @@ package ru.org.codingteam.keter.game.objects.equipment.items
 import ru.org.codingteam.keter.game.objects.equipment.{EquipmentItem, MeleeAttackCapability, Capability}
 
 abstract class Weapon extends EquipmentItem {
-  def damage: Double
+  def damage: Int
 }
 
 case class Knife(name:String,
                  actions: Set[Capability] = Set(MeleeAttackCapability),
-                 damage: Double = 100.0) extends Weapon
+                 damage: Int = 100) extends Weapon
